FIX: #49: New color codes for live run and exported image. FIX: #42: PNG image export for all runs of project. FIX: Number of Complete Runs in trial and various other fixes.
git-svn-id: https://codemax.unfuddle.com/svn/codemax_observador-svn@67 672ca433-b4e7-4185-b187-f88d0d116b9b

+++ src/observador/IEventVisualiser.cs
@@ -15,7 +15,7 @@ namespace observador
         void UpdateInterval(long milliseconds);
         void AddRunEvent(RunEvent runEvent);
         void SetBehaviors(List<Behavior> behaviors);
-        void SetDurationMilliseconds(int milliseconds);
+        void SetDurationMilliseconds(long milliseconds);
         void SetBehaviorColorAssigner(BehaviorColorAssigner behaviorColorAssigner);
     }
 }
